// src/app/shared/models/article.ts
import { contract } from "./contract";

export class article {
    id?: number;
    articleNumber?: number;
    articleTitle?: string;
    description?: string;
    contractId?: number;

}

export class articleUpdated {
    id?: number;
    articleNumber?: number;
    articleTitle?: string;
    description?: string;
    contractId?: number;article
}